Implement LDAP authentication
On the server side:

Introduced a new property called "http-server.authentication.type"
which can have values NONE, KERBEROS and LDAP.

You can have a custom user bind string property
(authentication.ldap.user-bind-pattern), which
will be used to construct the bind name to query LDAP.
This pattern *must* have a string '${USER}' as a placeholder
for the username and all instances of this string will
be replaced by the username.

Eg: "authentication.ldap.user-bind-pattern=${USER}@org.com"
with a username of 'test' will construct a bind name of "test@org.com".

Once the user's password is validated, the user's credentials will be
used to check if the user belongs to a particular group. The server
property 'authentication.ldap.group-auth-pattern' must have the group
check LDAP query.

On the CLI:
 --password flag will prompt for a password for the user