Fix hidden docs so they do not show up in the side bar
Keeping the docs in a hidden.rst preventing them from showing up on the main page index. But the docs still showed up in the side bar.
With this patch we add the docs to the exlude list. As a result some of the release notes needed updating because of missing references.