Verify window specification in plan and functions/frames in AST
The FunctionCalls in WindowNodes come directly from the AST, and are not filled
in with defaults when the user omits e.g. the orderings in a query. The DSL for
the unit tests should verify as much information about the window from the plan
and only the function calls and their associated frame information in the AST.

Conflicts:
	presto-main/src/test/java/com/facebook/presto/sql/planner/assertions/PlanMatchPattern.java

Conflicts:
	presto-main/src/test/java/com/facebook/presto/sql/planner/optimizations/TestMergeWindows.java

Conflicts:
	presto-main/src/test/java/com/facebook/presto/sql/planner/assertions/PlanMatchPattern.java