Fix incorrect stream property derivations from GroupIdNode
AddExchanges and AddLocalExchanges were inconsistent in the way that
properties were derived for the GroupIdNode. StreamPropertyDerivations
incorrectly passed through all child properties, when in fact the
GroupIdNode destroys partitioning properties on grouping columns that
are not common to all grouping sets. In some cases this led to
incorrect results. We now use the same logic in PropertyDerivations
and StreamPropertyDerivations.