Create GlobalProperties and pass to optimizers
Create GlobalProperties class that includes non-session properties that
are needed during planning.  Pass them through to the optimizers so that
they can be used during planning.  This is necessary for adding an
optimizer that can choose between distributed and broadcast joins based
on the size of the right-hand table as compared to
query.max-memory-per-node.