Add join_distribution_type session property
The join_distribution_type property has three option: "partitioned",
"replicated", and "automatic". This property is intended to replace the
"distributed_joins" property. "Partitioned" has the same behavior as
distribtued_joins=true. "Replicated" is equivalent to
distributed_joins=false. "Automatic" will use stats to evaluate whether
partitioned or replicated is better, or if no information is available,
it will choose partitioned.

The default value for join_distribution_type is "partitioned" (same as
for distributed_joins).  The distributed_joins property is
deprecated.  In the mean time, the distributed_joins and
join_distribution_type properties can conflict with each other.  The
expected behavior is as follows: if "join_distribution_type" property is
set to a non-default value, that value is used. If
"join_distribution_type" is set to a default value, we look at the
"distributed_joins" property.