Add TPC-DS queries as product tests
tpcds tests are disabled to be run in each travis build. It is going to
be run externally in prest-checks. This is due that execution duration
in about 2 hours.

Result files for quarantined could be not formatted properly to be
accepted by tempto. It is because, typically result files are generated
by tempto, but for queries which returns incorrect results result files
comes from TPC-DS benchmark.