Introduce subquery assignments to ApplyNode
This greatly simplifies apply node pruning
and rewrites of uncorrelated IN predicates.

Subquery expressions are different than other expressions
in a sense that they might use an entire subquery result
as an input (e.g: "x IN (subquery)", "x < ALL (subquery)").
Such expressions are invalid in linear operator context
(e.g: ProjectNode) in logical plan, but are correct in
ApplyNode context.

Example 1:
 - expression: input_symbol_X IN (subquery_symbol_Y)
 - meaning: if set consisting of all values for subquery_symbol_Y contains value represented by input_symbol_X

Example 2:
 - expression: input_symbol_X < ALL (subquery_symbol_Y)
 - meaning: if input_symbol_X is smaller than all subquery values represented by subquery_symbol_Y

Example 3:
 - expression: subquery_symbol_Y
 - meaning: subquery is scalar (might be enforced), therefore subquery_symbol_Y can be used directly in the rest of the plan

This changes the semantics of Apply from behaving as a cross join to a join against a scalar subquery.
Thus, it now preserves the cardinality of the input.