Follow docker logs
Starting from 1.7.0 docker composed does not follow the logs by default:
https://github.com/docker/compose/releases/tag/1.7.0

To make "logs" command to follow the logs "-f" parameters should be added.

This change is not compatible with the previous version of docker-compose.
Documentation updated that now "1.8.1" docker compose is required to run
product test. This is the version which is being used on Travis.