Add alias support for SetOperationNodes in DSL
SetOperationNodes rebind the symbols of their inputs. Those bindings
need to be passed to subsequent nodes in the plan tree. Additionally, a
single output symbol from the SetOperationNode is drawn from all of the
nodes input sources.  Therefore, we need to allow multiple aliases to
point to the same SymbolReference.