Fix building TupleDomain for comparisons with coercions
This patch fixes building TupleDomain for comparison expressions
where symbol-side of expression is wrapped with Cast

```
E.g. CAST(s AS BIGINT) >= 100000000000;
```

where s is INTEGER column

Algorithm is as follows:

We assume same expressions of same type on both sides of Comparison
expression.
First we do expression normalization which determines if
one side of comparison expression is foldable to literal and folding is
performed.
And that other side of comparison expression is either literal or
sequence of casts:
 CAST(CAST .... (CAST(symbol))))

During normalization we validate that all CASTS are coercions and bail
out if that is not the case.

Then.

If expression is of form:
```
  symbol OP literal
```
we build TupeleDomain of that.

If expression is of form:
```
  CAST(exp as T) OP literal
```

We rewrite expression to other shape which peels off one CAST and run
processing on rewritten expression. Saturated_floor_cast(litear, T) is
used in process of rewriting.