Replace exception with log.warn in RecordFileWriter.rollback()
FileWriter.rollback() is called in a loop and we want to call rollback
on other writers, even if one fails, so we do are not left with any
dangling resource. To achieve that in case of error in rollback() we
no longer throw exception out of it but just log a warning.